requirements.txt
pydantic
fastapi
uvicorn
requests
asyncio
tiktoken
python-dotenv
boto3
pandas
numpy
mistralai
openai
pinecone
snowflake-connector-python
snowflake-connector-python[pandas]
google-search-results
langchain-openai
langchain-community
langchain-core
langgraph
typing-extensions
yfinance
spacy
